+++ Service.Manafacturing.Model/Step/StepDetailModel.cs
@@ -11,6 +11,6 @@ namespace VErp.Services.Manafacturing.Model.Step
         public decimal Quantity { get; set; }
         public decimal WorkingHours { get; set; }
         public int NumberOfPerson { get; set; }
-        public decimal EstimateHandoverTime { get; set; }
+        public int EstimateHandoverTime { get; set; }
     }
 }
